requirements.txt
# Production RAG Chatbot Requirements

# Core LangChain
langchain==1.1.3
langchain-core==1.1.3
langchain-groq==1.1.0
langchain-huggingface==1.1.0
langchain_community==0.4.1
langchain_chroma==1.0.0

# LLM & Embeddings
sentence-transformers==5.1.2

# Document Processing
pypdf==6.4.1
docx2txt==0.9

# Vector Database
chromadb>=1.0.20

# FastAPI Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.9.2

# Streamlit Frontend
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
requests==2.32.5
